Migrate to asp.net core

+++ TaskManager.Database/dbo/Tables/Priorities.sql
@@ -1,5 +1,5 @@
 ﻿CREATE TABLE [dbo].[Priorities] (
-    [Id]   TINYINT        IDENTITY (1, 1) NOT NULL,
+    [Id]   INT        IDENTITY (1, 1) NOT NULL,
     [Name] NVARCHAR (100) NOT NULL,
     CONSTRAINT [PK_dbo.Priorities] PRIMARY KEY CLUSTERED ([Id] ASC)
 );
