+++ TaskManager.Database/dbo/Tables/States.sql
@@ -1,5 +1,5 @@
 ﻿CREATE TABLE [dbo].[States] (
-    [Id]   TINYINT       IDENTITY (1, 1) NOT NULL,
+    [Id]   INT       IDENTITY (1, 1) NOT NULL,
     [Name] NVARCHAR (50) NOT NULL,
     CONSTRAINT [PK_dbo.States] PRIMARY KEY CLUSTERED ([Id] ASC)
 );
